Refactor codebase: Clean up dumb-csv and reorganize ultros modules
- Remove abandoned `trash-heap` directory.
- Remove unused `add` function from `dumb-csv`.
- Refactor `ultros/src/main.rs`: move `run_socket_listener` to `socket_listener.rs` and `init_db` to `db_init.rs`.
- Refactor `ultros/src/web.rs`: move API handlers to `ultros/src/web/api/` modules (lists, retainers, characters, user, listings).
- Clean up unused imports.
