--- trash-heap/README.md
@@ -1,9 +1,0 @@
-# 🗑️ Trash Heap Info
-
-This project came about through many trial and errors. Originally the plan was to create an egui native desktop application, but eventually it became
-clear that making a web interface made more sense.
-
-### Repositories
-* `crafterstoolbox` - abandoned egui frontend that would price out all ingredients and give charts of demand
-* `recipepricecheck` - abandoned repo that wrapped universalis' api to figure out how much it'd cost to craft an item
-* `xivapi` - abandoned xivapi scraper, deprecated in favor of `xiv-gen`
