--- trash-heap/crafterstoolbox/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "crafterstoolbox"
-version = "0.1.0"
-edition = "2024"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-egui = "0.19"
-eframe = { version = "0.19.0", features = ["persistence"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-recipepricecheck = { path = "../recipepricecheck"}
-xiv-gen = {path = "../../xiv-gen", features = ["recipe", "recipe_lookup", "item", "item_ui_category"]}
-xiv-gen-db = {path = "../../xiv-gen-db"}
-universalis = {path = "../../universalis"}
-bincode = "2.0.0-rc.1"
-flate2 = "1.0.25"
-lazy_static = "1.4.0"
-futures = "0.3.23"
-anyhow = "1.0.62"
-serde-error = "0.1.2"
-fixed_decimal = "0.5.1"
-icu = {version = "1.0.0-beta1", features = ["serde"]}
-icu_testdata = "1.0.0-beta1"
-writeable = "0.5.0"
-log = "0.4.17"
-
-itertools = "0.10.3"
-chrono = "0.4.23"
-
-# native:
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tracing-subscriber = "0.3"
-tokio = {version = "1", features = ["full"]}
-
-# web:
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-console_error_panic_hook = "0.1.6"
-tracing-wasm = "0.2"
-tokio = {version = "1", default-features = false, features = ["sync"]}
