--- trash-heap/crafterstoolbox/src/lib.rs
@@ -1,35 +1,0 @@
-#![warn(clippy::all, rust_2018_idioms)]
-
-mod app;
-mod crafting_types;
-mod plots;
-mod sidepanel;
-// mod simulator;
-mod universalis;
-
-pub use crate::universalis::UniversalisData;
-pub use app::AppRx;
-pub use app::AppTx;
-pub use app::CraftersToolbox;
-
-// ----------------------------------------------------------------------------
-// When compiling for web:
-
-#[cfg(target_arch = "wasm32")]
-use eframe::wasm_bindgen::{self, prelude::*};
-
-/// This is the entry-point for all the web-assembly.
-/// This is called once from the HTML.
-/// It loads the app, installs some callbacks, then returns.
-/// You can add more callbacks like this if you want to call in to your code.
-#[cfg(target_arch = "wasm32")]
-#[wasm_bindgen]
-pub fn start(canvas_id: &str) -> Result<(), eframe::wasm_bindgen::JsValue> {
-    // Make sure panics are logged using `console.error`.
-    console_error_panic_hook::set_once();
-
-    // Redirect tracing to console.log and friends:
-    tracing_wasm::set_as_global_default();
-
-    eframe::start_web(canvas_id, Box::new(|cc| Box::new(CraftersToolbox::new(cc))))
-}
