--- trash-heap/recipepricecheck/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "recipepricecheck"
-version = "0.1.0"
-edition = "2024"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-universalis = {path = "../../universalis"}
-thiserror = "1.0.31"
-itertools = "0"
-serde = {version = "1", features = ["derive"]}
-xiv-gen = {path = "../../xiv-gen", features = ["recipe", "item"]}
-log = "0.4.17"
-
-[dev-dependencies]
-tokio = {version = "1.19.2", features = ["full"]}
-
-
-[lib]
-name = 'recipepricecheck'
