--- trash-heap/xivapi/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "xivapi"
-version = "0.1.0"
-edition = "2024"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-reqwest = {version = "0.11.10", default-features = false, features = ["rustls-tls", "json"]}
-serde = {version = "1.0.137", features = ["derive"]}
-serde-aux = "3.0.1"
-serde_json = "1.0.81"
-thiserror = "1.0.31"
-futures = "0.3.21"
-tokio = {version = "1", features = ["full"]}
-clap = {version = "3.2", features = ["derive"]}
-itertools = "0.10.3"
-log = "0.4.17"
