--- trash-heap/xivapi/src/models/mod.rs
@@ -1,4 +1,0 @@
-pub mod item;
-pub mod recipe;
-
-pub const BASE_URL: &str = "https://xivapi.com/";
